requirements.txt
django
djanogrestframework
djangorestframework-simplejwt
gunicorn
ipython
django-cors-headers